move conditional node operators to separate class
